Custom changes: fix problem calculations and loading init data

+++ examples/beam/beam.py
@@ -26,7 +26,7 @@ def main(job_id, params):
     c5 = 0.25 - delta
     
     # Combine constraints
-    c = float(c1 >= 0 and c2 >= 0 and c3 >= 0 and c4 >= 0 and c5 >= 0) - 0.5
+    c = (c1 >= 0 and c2 >= 0 and c3 >= 0 and c4 >= 0 and c5 >= 0).astype(float) - 0.5
 
     return {'f': f, 'c': c}
